pyproject.toml
[project]
name = "multi-source-rag-platform"
version = "0.5.1"
description = ""
authors = [
    {name = "Hagay Bar",email = "hagay.bar@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "python-pptx (>=1.0.2,<2.0.0)",
    "typer (>=0.16.0,<0.17.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "spacy (>=3.8.7,<4.0.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "litellm[proxy] (>=1.73.0,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "streamlit (>=1.46.1,<2.0.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "pylint (>=3.3.7,<4.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "colorlog (>=6.9.0,<7.0.0)"
    ]
   

[tool.ruff]
line-length = 100  # or lower for more aggressive wrapping

[tool.ruff.format]
docstring-code-format = true
quote-style = "preserve"

[tool.ruff.lint]
select = ["E", "F"]
ignore = [] 

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pylint = "^3.3.7"
ruff = "^0.12.5"

[tool.poetry.scripts]
litellm = "litellm.__main__:entrypoint"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
